cdk/utils/secure_vpc.py
'''
Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
Permission is hereby granted, free of charge, to any person obtaining a copy of this
software and associated documentation files (the "Software"), to deal in the Software
without restriction, including without limitation the rights to use, copy, modify,
merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
permit persons to whom the Software is furnished to do so.
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
'''

from aws_cdk import (
    aws_ec2,
    aws_kms
)
from constructs import Construct
from utils.vpc_flowlog import VpcFlowLog

class SecureVpc(aws_ec2.Vpc):

    def __init__(self, scope: Construct, id: str, kms_key: aws_kms.Key, **kwargs) -> None:
        super(SecureVpc, self).__init__(scope, id,  **kwargs)
        VpcFlowLog(scope, f"{id}FlowLog", 
            resource_type=aws_ec2.FlowLogResourceType.from_vpc(self), kms_key=kms_key)